+++ functions.php
@@ -7,7 +7,7 @@
 
 defined( 'ABSPATH' ) || exit;
 
-define( 'BLUE_THEME_VERSION', '2.9.3' );
+define( 'BLUE_THEME_VERSION', '2.9.4' );
 define( 'BLUE_THEME_DIR', get_template_directory() );
 define( 'BLUE_THEME_URI', get_template_directory_uri() );
 
@@ -20,6 +20,7 @@ require_once BLUE_THEME_DIR . '/inc/helpers.php';
 require_once BLUE_THEME_DIR . '/inc/media-assets.php';
 require_once BLUE_THEME_DIR . '/inc/setup.php';
 require_once BLUE_THEME_DIR . '/inc/acf.php';
+require_once BLUE_THEME_DIR . '/inc/smtp.php';
 
 // Storefront integrations.
 require_once BLUE_THEME_DIR . '/inc/woocommerce.php';